最近阅读增加双击打开
