+++ client-PC/renderer/components/book-list/index.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
+import utils from 'common/utils';
 import styles from './book-list.scss';
 import bookDefault from '../../../assets/images/book-default.png';
 
@@ -14,7 +15,7 @@ const mapDispatchToProps = () => ({
 });
 
 const bookRender = (book, idx) => (
-  <li key={`${book.rawname}-${idx}`}>
+  <li key={`${book.rawname}-${idx}`} onDoubleClick={utils.openFile(book.srcFullPath)}>
     <img className={styles.img} alt={book.rawname} src={book.cover && book.cover !== '' ? book.cover : bookDefault} />
     <p className={styles.name}>{book.titleDisplay}</p>
     <p className={styles.author}>{book.author && book.author.length > 0 ? book.author : 'Unknown Author'} </p>
